% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm
Finding greatest common divisor of two integers
By providing any two numbers and we can find highest common divisor for given
number. This function make things easy for us to
get highest divisor for numbers.}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{is any integer number}

\item{b}{is also an integer number}
}
\value{
This function returns a \code{number}
}
\description{
Euclidean algorithm
Finding greatest common divisor of two integers
By providing any two numbers and we can find highest common divisor for given
number. This function make things easy for us to
get highest divisor for numbers.
}
\examples{
euclidean(10,100)
euclidean(10,1000)

}
