% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Data for djisktra
Wiki_graph
that data will use for djisktra}
\format{
A data frame with three enrtries:
\describe{
\item{v1}{contain data for first node}
\item{v2}{contain data for second node}
\item{w}{contain data for paths between nodes}
}
}
\usage{
wiki_graph
}
\description{
Data for djisktra
Wiki_graph
that data will use for djisktra
}
\keyword{datasets}
